align with options structure
