+++ source/IntegrationEvents.Application/IReceivedIntegrationEventRepository.cs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Data;
-using System.Threading.Tasks;
 
-namespace Energinet.DataHub.EDI.IntegrationEvents.Infrastructure;
+namespace Energinet.DataHub.EDI.IntegrationEvents.Application;
 
 /// <summary>
 /// Persists received integration event metadata
