+++ source/IntegrationEvents.Infrastructure/IntegrationEvents.Infrastructure.csproj
@@ -8,6 +8,8 @@
       <ProjectReference Include="..\AuditLog\AuditLog.csproj" />
       <ProjectReference Include="..\BuildingBlocks.Infrastructure\BuildingBlocks.Infrastructure.csproj" />
       <ProjectReference Include="..\Contracts\Contracts.csproj" />
+      <ProjectReference Include="..\DataAccess\DataAccess.csproj" />
+      <ProjectReference Include="..\IntegrationEvents.Application\IntegrationEvents.Application.csproj" />
 
       <ProjectReference Include="..\MasterData.Interfaces\MasterData.Interfaces.csproj" />
     </ItemGroup>
